karepalAI/requirements.txt

beautifulsoup4==4.12.3
binance==0.3
binance-futures==0.0.1
binance-futures-connector==4.0.0
Flask==3.0.3
Flask-RESTful==0.3.10
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.1
google_images_download==2.8.0
googleapis-common-protos==1.63.0
joblib==1.4.2
gTTS==2.5.1
gunicorn==22.0.0
mysql-connector-python==8.3.0
pandas==2.2.2
numpy==1.26.4
oauthlib==3.2.2
mysql-connector-python==8.3.0
oauthlib==3.2.2

pickle-mixin==1.0.2
pillow==10.3.0

PyMySQL==1.1.0
pyOpenSSL==24.1.0

pyperclip==1.8.2

pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
python-binance==1.0.19
requests==2.31.0
requests-oauthlib==2.0.0

selenium==4.19.0
Flask-Cors==4.0.1
SpeechRecognition==3.10.3
sqlparse==0.4.4

urllib3==2.2.1
